# Conflicts:
#	apps/exampleViewer/ospExampleViewer.cpp
#	apps/exampleViewer/widgets/imgui3D.cpp

+++ apps/bench/bench.cpp
@@ -50,8 +50,8 @@ namespace ospray {
 
   std::vector<std::string> files;
   std::string imageOutputFile = "";
-  int numWarmupFrames = 10;
-  int numBenchFrames  = 100;
+  size_t numWarmupFrames = 10;
+  size_t numBenchFrames  = 100;
   int width  = 1024;
   int height = 1024;
